user-service/requirements.txt
flask
flask-sqlalchemy
python-dotenv
bcrypt
PyJWT
flask-migrate
psycopg2-binary